[fix-cdburnerxp] fix download link and essential changes:
Since CDBurnerXP change its download link of minimal package in very short time, it is impossible to track the change in simple way.
The only choice is to use its full package instead.

In options.%lang%.cfg files, "SetupType" changes to "custom". It is to avoid risk of installing useless component.

May be go for minimal package again in future, but I need to call for help from developer of CDBurnerXP.